Refactored module defs
