+++ external-crates/move/crates/move-analyzer/src/completions/name_chain.rs
@@ -20,7 +20,8 @@ use crate::{
     symbols::{
         cursor::{ChainCompletionKind, ChainInfo, CursorContext},
         def_info::{DefInfo, FunType, VariantInfo},
-        types::{AutoImportInsertionInfo, MemberDef, MemberDefInfo, ModuleDefs, Symbols},
+        mod_defs::{AutoImportInsertionInfo, MemberDef, MemberDefInfo, ModuleDefs},
+        types::Symbols,
     },
     utils::expansion_mod_ident_to_map_key,
 };
